[CL-3636] Move new code to new ExamplesService
